ajustando README.md e configurações para executar no postgreSQL

+++ academia/models.py
@@ -23,13 +23,13 @@ class Cliente(db.Model):
     nome = db.Column(db.String(20), nullable=False)
     sobrenome = db.Column(db.String(50), nullable=False)
     genero = db.Column(db.String(1), nullable=False)
-    cpf = db.Column(db.Integer, nullable=False, unique=True)
-    rg = db.Column(db.Integer, nullable=False, unique=True)
+    cpf = db.Column(db.BigInteger, nullable=False, unique=True)
+    rg = db.Column(db.BigInteger, nullable=False, unique=True)
     dt_nascimento = db.Column(db.DateTime, nullable=False)
     estado_civil = db.Column(db.String(20), nullable=False)
 
     email = db.Column(db.String(100), nullable=False, unique=True)
-    telefone = db.Column(db.Integer, nullable=False)
+    telefone = db.Column(db.BigInteger, nullable=False)
 
     rua = db.Column(db.String(100), nullable=False)
     numero = db.Column(db.Integer)
